+++ src/pages/main-page/main-page.test.tsx
@@ -1,12 +1,13 @@
 import { render } from '@testing-library/react';
+import { IMessage } from '../../components/message/message-interfaces';
 import Main from './main-page';
 
-const mockSetError = vi.fn((message: string) => {
-  const state: { errors: string[] } = {
+const mockSetError = (message: IMessage) => {
+  const state: { errors: IMessage[] } = {
     errors: [],
   };
   state.errors.push(message);
-});
+};
 
 test('Main page is load', () => {
   const wrapper = render(<Main setError={mockSetError} />);
